requirements.txt

streamlit==1.29.0
pandas==2.1.1
langchain==0.0.308
langchain-openai==0.0.2.post1
langchain-community==0.0.13
openai==1.3.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2